Common get path function
